Fix the build with pango main
A new attribute type has been added. Unfortunately,
that breaks the build in some places.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/10086>

+++ demos/icon-editor/fontify.c
@@ -172,7 +172,7 @@ insert_tags_for_attributes (GtkTextBuffer     *buffer,
     {
       PangoAttribute *attr = l->data;
 
-      switch (attr->klass->type)
+      switch ((int) attr->klass->type)
         {
         case PANGO_ATTR_LANGUAGE:
           LANGUAGE_ATTR (language);
